Move slow tests (>= 10 seconds) to the nightly check, fixes #606
Note: category annotations only take effect when on the class.
On my first pass of this ticket, I marked the test methods instead of the classes.
I think it makes sense to leave the annotations on the methods as they provide
some context on the slow methods for the given class.