JS API Error handling (#1086)
Added basic error handling, but not proper reconnects or reauth. Three error types are handled for the main streams or polling calls:

    auth failure which suggests either that the server was restarted or the session timed out somehow - in either case, a new auth is required to resume working
    Code.INTERNAL, which our grpc-web client library will pass in cases where the response doesn't make sense (broken payloads)
    Code.UNKNOWN, which our grpc-web client library will pass in cases where the transport doesn't seem to work (no response before connection died)

A new event was added, temporarily, to inform the UI that we are giving up if one of these happens, and the page needs to be reloaded. This will be removed when we have proper reconnect, and inform the UI again that the reconnect succeeded or failed.

Also modified a few errors being sent from the server to hopefully better reflect the intent of the gRPC status codes, and avoid using "INTERNAL" when "something bad" happens.

The UI will need a few small changes to handle this, and to handle a few cases where API calls can fail but not be registered in the UI.

Fixes #1065, but #730 needs to formally address this.