Docker tasks should always tell gradle if they have changes
Previously, if there was no copyOut, a docker task would not declare any
outputs, so downstream tasks would not know if they had to re-run. This
change explicitly marks the built image as the output, and logs a
message in case an entrypoint running despite its output being ignored.

Fixes #1408