[jsapi] Set rollback to null to enable orphan cleanup (#1476)
Previously, rollbacks were not nulled out, in an attempt to avoid losing useful information as the table transitioned between states. Bug reports that applied to these old changes were used to test this commit; this does not alter their behavior (though I could not test the dashboard related bug).

Now, when an operation reaches a terminal state, its rollback should be cleared so that it is eligible for cleanup if it ever becomes an orphan.

While testing I also came across an additional CTS assertion error. Failed requests must not attempt to set their client state as running. If a table reaches a terminal state, it must always have its rollback nulled out, to avoid leaking the rollback.

Minor test change to table_basic.html to use global variables so that they are easily accessible from the developer console.

Fixes #1471