Reduce scope of dependencies for faster/simplier rebuilds (#1707)
The "new" dependencies introduced in this commit were ones that were transitive before with no need to be (i.e. not part of the actual API, but exposed anyway). Some of these are negotiable, like what should the grpc-api project actually expose to downstream projects - this is just a proposal that appears to build as-is, cherry-picked from the jetty work, to make that eventual merge smaller.

Partial #1270