Avoid removing session close callbacks if session already expired
This should eliminate a ConcurrentModificationException for some clients
that have enough objects in their sessions in the right order.

Also added a bit more documentation to clearly acknowledge that while
Closeable.close() is expected to be idempotent, the SessionState will
not call it more than once, and provides removeOnCloseCallback to help
downstream code make its own guarantees about this.

Fixes #1651