Add numeric conversion in DynamicTalbeWriter.java
Given the choice, users will be drawn to use the convenience methods
logRow() in DynamicTableWriter which accept generic java.lang.Object
in the input parameters. So there is a chance the numeric values don't
match the declared Column data type are passed in. As performance isn't
really a big concern here, we can afford to perform conversions and make
it easier for the users than require them to do cast/conversion before
calling the methods.