Exclude certain engine methods from Tier 4 compilation (#1985)
This commit excludes 60+ methods from hotspot's Tier 4 (aka C2) JIT
compilation. Each of these has been found crash the certain JVM
implementations (across vendors, across OSes, and across hardware
archetectures) when it compiles this code, but not consistently. By
excluding only C2 in this way, we allow C1 (aka Tier 1-3) to still run,
and get better performance than interpreted bytecode.

All existing ways to run code should be affected by this - running
main()s or tests through either gradle or IntelliJ, and running our
shell scripts to launch the server will all end up with a compiler
directives file to disable c2 for these methods. Using the engine as a
library is not fixed in this way, but this is a problem that can only be
solved through documentation, to inform users that they should make sure
to start the JVM with a specific set of flags.

An extra array of "devJvmArgs" is left in the io.deephaven.java-conventions
plugin to make it easier to add more flags later for this kind of
problem - it will automatically be included for all of the above categories.

Fixes #1529