Python library for embedded server (#2353)
Introduces a Python library that can start the Deephaven server from
python code. This consists of a wheel that holds not only some python
glue to start the server, but all of the jars required to be on the
deephaven classpath. Presently this depends on deephaven's own jpy
fork, but soon this limitation will be dropped and the upstream jpy
project will work.

Partial #2221