Switch to Parquet-specific snappy codec, and limit input stream size (#2562)
This restores the native parquet codec for snappy, but with a version
that supports aarch64 on macs. The other snappy codec was for hadoop,
not for parquet, and apparently there are differences between those
two.

The input stream size must be set correctly - The earlier patch was
written under the mistaken impression that the "size" field of the
IOUtils.buffer() method would set a max size to buffer. Not all codecs
fail with this error, since some stop decoding or are able to decode
other payloads in the parquet file, but this change at least will
speed up those implementations too.

Fixes #2495