Issue #2570 - Special variables in query language use type from queryscope variables
The special keywords i, ii, key, etc., are meant to be reserved and normally
they can't be used in the QueryScope. They are blocked by, e.g.,
QueryScope.StandaloneImpl.putParam. However in the Python and Groovy context,
they can effectively slip into the QueryScope because the user may set them
in the underlying session (e.g. by saying i = 12) and then
SynchronizedScriptSessionQueryScope or UnsynchronizedScriptSessionQueryScope
will just pass them through to the caller. With this change, the above classes
filter out the reserved names so the caller does not see them. From the user's
point of view, it is as if the special keywords i, ii, key, etc. have an inner
scope that hides any outer definition of those names.