Go Client API (#2497)
This is the initial revision of the Go client library for Deephaven.
Using the gRPC/Flight api, this can start a session, create tables
and run remote code, and fetch data. Ticking deltas are not yet
implemented, but data is instead read and written using standard
Apache Arrow Flight operations. 

Fixes #2523

Co-authored-by: Nathaniel Bauernfeind <natebauernfeind@deephaven.io>
Co-authored-by: Chip Kent <chipkent@deephaven.io>
Co-authored-by: Colin Alworth <colinalworth@deephaven.io>