ExecutionContext: record state of globals before deferring work (#2539)
The QueryScope, QueryLibrary, CompilerTools were merged into a new package `io.deephaven.engine.context` and are bundled as a distinct artifact `engine-context`. By default the ThreadLocal is "poisoned" to prevent accidental use of this state in unexpected contexts. The session's execution context will be installed from all entry points to the gRPC API as well as scripts run in a deephaven script session. To use elsewhere it is encouraged to capture the least state needed and pass the new ExecutionContext to use on another thread or during a deferred task. 

WhereFilters and SelectColumns will be initialized immediately and copies will retain details instead of needing to reevaluate in deferred contexts.

This commit introduces `TableOperationsDefaults` which lifts many default `TableOperations` implementations from `TableDefaults`. This helps reinforce consistency in other implementations of TableOperations.

The QueryLanguageParser can now deduce the return type of a generic method if it is fixed due to inheritance rules (both via interfaces and parent classes).