Implement gRPC storage API for notebooks and layouts (#2843)
This patch aims to replace existing WebDAV use cases with something that can be rolled into the main server process, and leaves room to establish additional authorization rules and use case conventions around this feature. The present implementation is built to allow the server implementation to be replaced by something else that would use deephaven-server dependencies and is backed by something other than the file system.

The expectation is that this isn't specific to notebooks, but serves as a general storage API for clients, so they can read/write their own data/code, and possibly share some code. Likely, this would be read-only or largely unavailable by non-privileged users. It could be that different users see different contents to the same top-level directory in some future implementation, but for this version a single directory is made available to all users.

Also introduces a JavaScript API for this and other upcoming features, in the form of dh.CoreClient, a starting point to connect to a deephaven-core instance or cluster. At this time authentication is not implemented, but will be soon - only the storage service is exposed. 

Fixes #1768