Use one websocket per browser client (#2606)
Creates a new websocket protocol to let the websocket behave only as a
transport on which multiple streams, across services, can be conveyed.
The websocket will be opened to the path matching the first stream, but
after the websocket has started, other streams can be sent on it, using
custom metadata/header to give the path to the desired rpc.

By reusing existing paths, the client can signal that it supports
either old or new websockets, so it in theory can connect to the old
grpc-websockets implementation. Unfortunately, the reference
implementation can't handle alternative requested subprotocols, and
will just fail the entire request, so this patch breaks our
compatibility with that proxy, and so with our netty implementation.

On the other hand, the server will maintain compatibility for both
protocols for now, enabling stale js api clients to still be able to
connect. We may also find that there are issues with longlived sockets,
and might end up wanting to ask the client in some circumstances to use
the old implementation.

Fixes #2510