Wrap Jetty's Resource type to force an invalid last-modified header (#3045)
Jetty always returns the last-modified date of the file, but Gradle
unsets the "true" last modified date in favor of ensuring that build
files cache only based on their contents. Disabling this feature is
possible, but still may present as buggy in Jetty, such as if a 1.0
release occurred, then 2.0, then a 1.1 hotfix - if an installation of
1.1 were later replaced with 2.0, the 1.1 files would be "newer" so
would be considered up to date.

This patch removes the last-modified value by only returning -1 when
requested. We could go further and return an etag that actually maps to
the release version or the contents of the file, but that is left for
future work.

The patch also removes the If-Modified-Since header from incoming
requests, so that Jetty doesn't attempt to tell the client it has an up
to date copy already. In the future, if we add an etag implementation
that we're happy with, we might remove this header, but for now if this
header is allowed, we risk updating users getting stuck with stale
files.

Fixes #3011