Refactor dagger mains for better extensibility (#3159)
This PR moves JettyMain from server-jetty to server-jetty-app; the main class isn't really meant to be a "library". That said, it's a very few short lines of code for integrators to achieve the same thing that JettyMain does:

```java
public final class MyCustomMain {
    public static void main(String[] args)
            throws IOException, InterruptedException, ClassNotFoundException, TimeoutException {
        final Configuration configuration = MainHelper.init(args, MyCustomMain.class);
        new CommunityComponentFactory()
                .build(configuration)
                .getServer()
                .run()
                .join();
    }
}
```

This PR also uses the "Community" prefix in more places to hint to integrators that it may or may not be appropriate for "Enterprise" contexts.

This PR also introduces server-jetty-app-custom as an example for how an integrator can customize the structure and configuration of the Deephaven server. It's also useful from the perspective of dogfooding and ensuring DHC developers are aware of any changes that may break integrators.

As part of the above, ComponentFactoryBase has been added as an abstraction around creating Components. While not technically necessary, it's a useful construct that allows constructing components without callers needing to directly invoke dagger.