Update server-base and server-slim-base to production-like images.
This was a longer-than-expected effort to close the gap between production images, and integration testing images. The big updates: the base images are now based on ubuntu:22.04 (instead of an old debian), and the python version is now 3.10 (instead of an old 3.7). These images are now being produced with Dockerfiles from https://github.com/deephaven/deephaven-server-docker that are as similar to the production images as possible.

Necessary testing fixes already done in #3331