Add extraHeaders support for java-client
This adds the ability for construction of a java-client with extra headers set on every request. This is potentially useful for cases where a proxy needs additional information to be able to route a request. It has been plumbed through the java client example utilities. In the future, we may wish to more formally support these use cases through DeephavenTarget / DH URIs.