JS API should attempt to reconnect to the gRPC server (#3502)
As long as a user's session on the server remains active, this should
enable a client to reconnect to that session, and continue using
already-exported objects. Some of the work is restored here for #3501
to be finished as well, but the client would need a way to create a new
session without user interaction.

The dh.IdeConnection.HACK_CONNECTION_FAILURE event is now deprecated,
and should be removed. Clients should instead use the
EVENT_DISCONNECTED and EVENT_RECONNECTED events to signal to the user
that an object isn't available. Alternatively, to detect shutdown, use
EVENT_SHUTDOWN.

Also fixes a bug where the last seen log item is replayed incorrectly.

Fixes #730
Fixes #225