KafkaTools implementation refactor.
This is a non-trivial refactoring of the KafkaTools from a bunch of switch statements into abstract methods. This should make existing implementations easier to maintain, and new implementations easier to implement. Furthermore, this allows implementations to construct explicit Serializer/Deserializers, as well constructing SchemaRegistryClient once - this is a pre-requisite for protobuf support.