Fix thread safety for gRPC websocket transports  (#4408)
Some gRPC operations on ServerTransportListener and TransportState are
specified as only intended to be called from the "transport thread" -
prior to this patch, we had incorrectly interpreted this to allow
serial operations from the websocket handler in the server. This patch
ensures that those operations are additionally serialized against
calls from gRPC internals. Among other things, this fixes an issue
where messages could be closed while being deframed.

Fixes #4406