Fix kafka json byte and char types
byte consumption was incorrect; it was being deserialized as a text node with the first byte of that text being the result. Instead, it should be treated as a numeric node, and cast to byte. This matches how other numeric types are handled.

char production was incorrect; it was being serialized as an int node. Instead, it should be treated as a text node which matches the expectations of the existing deserialization logic. Additionally, we can extend char deserialization to work in the case where the node actually _is_ a number.

Fixes #4384