Enhance Python LivenessScope API (#4497)
The existing Python APIs for Liveness were difficult to use correctly
in cases more complex than a single script - this brings several parts
of the Java LivenessScope type to Python to enable more advanced use
cases. Rather than introduce a LivenessScopeStack, this adds an open()
method to LivenessScope itself. The liveness_scope() method will
continue to exist for simple use cases, where the scope will be closed
and released at the same time at the end of a block or function.

Also makes TableListenerHandle a JObjectWrapper so it can be more
easily unwrapped, can't accidentally be managed as a
LivePyObjectWrapper in Java.

Fixes #4471