Add KafkaTools.Consume#objectProcessorSpec for custom KeyOrValueSpecs
This introduces a new public interface, ObjectProcessor, that provides a means for end-users to create custom KeyOrValueSpec for use with Kafka ingestion. The interface isn't specific to kafka though, and should be extensible to other domains in the future.

The existing KeyOrValue specs are left unchanged; but arguably, there may be value in porting our existing implementations over to it in the future.

Fixes #4346