Enhance Groovy classloader experience (#4596)
Introduces new features when using Groovy as a console language. This
should not remove or change existing features at all, but only serves
to backfill the removed source() and sourceOnce() functions with an
approach that encourages users to write real groovy scripts.

Makes it possible to load .groovy files from the classpath (or a custom
classloader) as classes. Standalone groovy scripts provide a static
main() method (and an instance run() method) to invoke one or more
times, and classes can be reimported to read fresh from disk. Includes
a feature where groovy files loaded from the classpath can use
Deephaven default imports, disabled by default.

Replaces a list of import statements with the Groovy ImportCustomizer,
so that evaluated code doesn't have a variable number of prefix lines.
This has the advantage of also being applicable to scripts loaded from
the classpath.

Restored unit tests lost when migrating to the deephaven-core codebase,
which helped to validate some of these cases, and verify
console<->querylanguage classloader interactions.