DoExchange snapshot won't half-close until after client does
Additional mitigation for
https://github.com/envoyproxy/envoy/issues/30149 - combined with a
client that will always half-close before the server does, this ensures
that the server will never need to half-close first in order to end a
stream. This is technically a breaking change - existing clients which
fail to half-close will instead hang with the connection left open until
timeout.

Fixes #4627