Fix intermittent build failure, ensuring docker test output is still usable (#4692)
This PR reverts #4254, and rewrites the changes to now only support the
use case that it was written for, but also avoid the gradle artifact
issue that it appears to have introduced. The patch works by both using
dependsOn and finalizedBy to set up two different downstream Sync tasks
that do the same thing - but only one runs on success, and only one
runs on failure. 

Fixes #4691