Added in an experimental Windowed table.
This implements a windowing partitioned table, that can take in timestamped data and window
across any period of time for any amount of backlook desired.

Generally the implementation leverages an input Blink table that has a timestamp column, it
then creates a new __WINDOWED__ column that it partitions the input by. The implementation
leverages a PriorityQueue to then clean up "expired" windows, maintaining the specified number
of windows and correctly partitioning the incoming data across those windows.