Extract variable api from ScriptSession, let ScriptSession guard reads
This patch moves from assuming that the ScriptSession owns the
ExecContext for locking (with UpdateGraph) and variable access (with
QueryScope).

Previously, the groovy script session used a lock to interact with its
query scope, but python did not under the incorrect assumption that the
GIL would guard those reads. Instead, a read/write lock is used to
ensure that reads block while an operation is running that could affect
values, but avoiding reads blocking each other.

Partial #4040