Update publish-ci to continue on certain failures
There was an error observed during the v0.33.0 release where the deephaven-server artifact was too large for PyPi. There are other failure scenarios that we could potentially run into here as well - network connectivity, expired tokens, temporary service interruptions, etc. Given that all of the release artifacts have already been successfully uploaded as a GitHub artifact, we should prefer to continue with publishing to all the sources even if one of the sources fails.