Adds a quick CI job to verify python minimum version
We've had a few recent issue/PRs dealing with python minimum version support: #5227, #5235, and #5271

Ultimately, we'd like a full matrix of testing support: #3724, #3725

In the more immediate term, there is a tool that I've verified is capable of catching these sorts of issues: https://github.com/netromdk/vermin. I've verified that it does catch #5227 and #5271, but does not catch #5235 (which seems like it is more of a runtime error).

This PR adds a quick CI job to verify a python minimum version of 3.8. The `# novermin` comment is necessary in a few locations where we've explicitly / manually worked around python minimum version support.