Add better configuration objects for Session, FlightSession, and BarrageSession
This allows for the cleaner separtion between connection (ClientConfig) concerns and session concerns (SessionConfig, FlightSessionConfig, BarrageSessionConfig), as well as removing the dagger fluff that is not adding value. It will now be possible to use a single ManagedConnection and to create separately authenticated Sessions, something that was not easily possible before.