Fix WhereFilterAdapter for String literals (#5408)
It was not obvious during the initial implementation, but `RangeConditionFilter` expects String literals to be wrapped in quotes (as it primarily is sourced from `WhereFilterFactory` / query parsing logic). It was also discovered that one of the `MatchFilter` constructors is very non-obvious, and can lead to case-insensitive String matching. This was the case not only in `WhereFilterAdapter`, but also in `PartitionedTable#constituentFor`. The `MatchFilter` constructor has been marked as deprecated, and all internal usages of it have been migrated explicitly. This may be a good motivator to implement #3730.

Fixes #5407