Publish deephaven-server-test-utils jar (#5467)
Ultimately, this PR makes the `server-test-utils` jar public so external integrators can write unit tests against it. As part of this effort, all transitive dependencies of `server-test-utils` need to be made public as well. test-configs was elevated to a public jar, and in other cases, "test library code" was extracted from "test code" (in these cases, we already had appropriate public testing jars the code could be moved to).

Fixes #5341