Fix ring table for nested arrays
This reduces the type checking for ring table chunk sources; essentially, matching the expectations that we might need to copy around empty `Object[0]` as a substitute for the more specifically typed empty array. While I'm not the biggest fan of persisting this (ab)use of the type system, it is out of scope to actually try and plumb correct typing throughout the stack at this time.

Fixes #5498