Add java-client support for primitive array types (#5494)
This is a quick pass at adding support for sending primitive typed arrays, String, and Instant[] with the java client. A more thorough (and dense) recursive solution was initially created, but this was a more explicit version created for expediency.

The better solution can come at a later date. Ideally, we'd also add integration tests. In lieu of that, a new example utility that demonstrates all of the types was created.

Fixes #4721