Add defensive bearer token rotation for java client
This solves a theoretical flashing issue that could arise during token rotation where, depending on the order of responses from the server, a bearer token could be rotated to the new value and then "rotated" back to the old token value. While this situation does not cause an issue with our current server implementation (which is rather lenient and keeps around multiple previous bearer tokens), we should still prefer to be defensive on the java client side and use the newest value.