Update to GWT 2.11, clean up old gradle wiring (#5576)
Precursor to changes for #188 - updating to GWT 2.11 will make it
smoother to follow up with additional updates. The GWT 2.10 release
changed the groupId from com.google.gwt to org.gwtproject, but our
Gradle plugin does not support the new groupId, so this patch also
rewrites automatic dependencies added by the plugin to use the new
version.